requirements.txt
dash>=2.16.0
plotly>=5.20.0
pandas>=2.0.0
python-dotenv>=1.0.0
ctrader-open-api>=0.9.0
twisted>=24.3.0
service-identity>=21.1.0
pytest>=7.0.0
